requirements.txt
black==21.9b0
Django~=3.2.7
djangorestframework==3.12.2
djangorestframework-simplejwt==4.8.0
djangorestframework-filters==0.10.2
gevent==21.1.2
gunicorn==20.0.4
Pillow==8.3.2
pre-commit==2.12.1
pylint==2.8.2
pylint-django==2.4.4
python-decouple==3.4
psycopg2==2.9.1
requests==2.20.0
